+++ src/components/service.js
@@ -111,7 +111,7 @@ export class Service extends Component {
 	async create(obj) {
 		// if obj is a plain object, create a new entity
 		let entityInstance =
-			(!obj) instanceof this.entityClass ? this.format(obj) : obj;
+			obj instanceof this.entityClass ? obj : this.format(obj);
 
 		await this.remote
 			.invoke(this.remoteMethods.create, entityInstance)
@@ -148,7 +148,9 @@ export class Service extends Component {
 	}
 
 	async update(obj) {
-		let entityInstance = this.format(obj);
+		let entityInstance =
+			obj instanceof this.entityClass ? obj : this.format(obj);
+
 		await this.remote
 			.invoke(this.remoteMethods.update, obj)
 			.then((response) => response.json())